backend/services/model_comparison.py
# File: backend/services/model_comparison.py
"""
Model comparison module using cross-validation.
Determines which model has the highest accuracy.
"""
from sklearn.model_selection import cross_val_score
import pandas as pd

class ModelComparator:
    """Compares different model performances."""
    
    def __init__(self):
        self.results = {}
    
    def compare_models(self, X: pd.DataFrame, y: pd.Series, models: dict) -> dict:
        """Performs cross-validation on all models and compares accuracy."""
        # TODO: Implement model comparison logic
        pass
    
    def get_best_model(self) -> str:
        """Returns the name of the best performing model."""
        # TODO: Implement best model selection
        pass
