backend/services/report_generation.py
# File: backend/services/report_generation.py
"""
AI-powered report generation using OpenAI API.
Creates business insights based on analysis results.
"""
import os
from dotenv import load_dotenv
import openai

class ReportGenerator:
    """Generates business insights reports using AI."""
    
    def __init__(self):
        load_dotenv()
        self.api_key = os.getenv("OPENAI_API_KEY")
        # TODO: Configure OpenAI client
    
    def generate_report(self, forecast_data: dict, products_data: dict) -> str:
        """Generates comprehensive report based on analysis results."""
        # TODO: Implement report generation logic
        pass