requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
scikit-learn==1.4.0
pydantic==2.6.1
python-dotenv==1.0.0
numpy==1.26.3
openai==1.12.0
python-multipart==0.0.9